requirements.txt
annotated-types==0.7.0
ansi2html==1.9.1
attrs==23.2.0
blinker==1.8.2
Brotli==1.1.0
cachelib==0.9.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
dash==2.17.1
dash-bootstrap-components==1.4.2
dash-core-components==2.0.0
dash-extensions==1.0.16
dash-html-components==2.0.0
dash-table==5.0.0
dash_renderer==0.13.0
dashtable==1.4.5
dataclass-wizard==0.22.3
distro==1.9.0
django-wkhtmltopdf==3.4.0
EditorConfig==0.12.4
et-xmlfile==1.1.0
Flask==2.2.5
Flask-Caching==2.3.0
Flask-Compress==1.15
Flask-Mail==0.10.0
fonttools==4.53.0
h11==0.14.0
html5lib==1.1
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsbeautifier==1.15.1
MarkupSafe==2.1.5
more-itertools==10.3.0
mysql-connector-python==8.4.0
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pango==0.0.1
pdfkit==1.0.0
pillow==10.3.0
plotly==5.22.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydyf==0.10.0
pyodbc==5.1.0
pypdf==4.2.0
pyphen==0.15.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.2.0
requests==2.32.2
retrying==1.3.4
ruff==0.4.9
selenium==4.21.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tabula-py==2.9.3
tabulate==0.9.0
tenacity==8.3.0
tinycss2==1.3.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
weasyprint==62.2
webdriver-manager==4.0.1
webencodings==0.5.1
Werkzeug==2.2.3
wkhtmltopdf==0.2
wsproto==1.2.0
zipp==3.19.2
zopfli==0.2.3
zstandard==0.22.0
